+++ R/raw-conversion.R
@@ -1,7 +1,7 @@
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Convert a raw vector into a ARGB32 pointer
+#' Convert a raw vector into a packed color data pointer
 #'
 #' This function copies the original data into a new object and returns a
 #' pointer to this data.
@@ -9,25 +9,25 @@
 #' @param raw_vec raw vector
 #' @param width,height image dimensions
 #'
-#' @return An external poitner to ARGB32 data
+#' @return An external pointer to packed pixel data
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-raw_to_argb32 <- function(raw_vec, width, height) {
-  .Call(raw_to_argb32_, raw_vec, width, height)
+raw_to_packed <- function(raw_vec, width, height) {
+  .Call(raw_to_packed_, raw_vec, width, height)
 }
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Convert an ARGB32 pointer to a raw vector
+#' Convert a packed data pointer to a raw vector
 #'
-#' This function copies the data in the ARGB32 pointer to a new raw vector and
+#' This function copies the data in the packed pointer to a new raw vector and
 #' returns the new raw vector
 #'
-#' @param argb32_ptr ARGB32 pointer
+#' @param packed_ptr pointer to data containing packed colors
 #'
 #' @return raw vector
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-argb32_to_raw <- function(argb32_ptr) {
-  .Call(argb32_to_raw_, argb32_ptr)
+packed_to_raw <- function(packed_ptr) {
+  .Call(packed_to_raw_, packed_ptr)
 }