+++ src/aaa.h
@@ -1,0 +1,7 @@
+
+
+
+
+#define ARGB32 0
+#define RGBA32 1
+#define ABGR32 2