+++ src/raw-conversion.c
@@ -13,16 +13,16 @@
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Copy the data in a raw vector to newly allocated memory for ARGB32 data
+// Copy the data in a raw vector to newly allocated memory for packed data
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP raw_to_argb32_(SEXP raw_, SEXP w_, SEXP h_) {
+SEXP raw_to_packed_(SEXP raw_, SEXP w_, SEXP h_) {
 
   int w = INTEGER(w_)[0];
   int h = INTEGER(h_)[0];
   int l = length(raw_);
 
   if (l != w * h * 4) {
-    error("raw_to_argb32_(): Length of raw vector (%i) does not match w*h*4 = %i", l, w*h*4);
+    error("raw_to_packed_(): Length of raw vector (%i) does not match w*h*4 = %i", l, w*h*4);
   }
 
   unsigned char *dst = (unsigned char *)calloc(w*h*4, sizeof(unsigned char));
@@ -44,16 +44,16 @@ SEXP raw_to_argb32_(SEXP raw_, SEXP w_, SEXP h_) {
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Copy ARGB32 data into a raw vector
+// Copy packed data into a raw vector
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP argb32_to_raw_(SEXP argb32_) {
+SEXP packed_to_raw_(SEXP packed_ptr_) {
 
-  unsigned char *argb32;
+  unsigned char *packed_ptr;
 
-  argb32 = isNull(argb32_) ? NULL : (unsigned char *)R_ExternalPtrAddr(argb32_);
-  if (argb32 == NULL) error("argb32_to_raw_: 'unsigned char * src' pointer is invalid/NULL");
+  packed_ptr = isNull(packed_ptr_) ? NULL : (unsigned char *)R_ExternalPtrAddr(packed_ptr_);
+  if (packed_ptr == NULL) error("packed_ptr_to_raw_: 'unsigned char * src' pointer is invalid/NULL");
 
-  SEXP prot_ = R_ExternalPtrProtected(argb32_);
+  SEXP prot_ = R_ExternalPtrProtected(packed_ptr_);
   if (isNull(prot_)) {
     error("Need 'prot' set on external pointer");
   }
@@ -62,7 +62,7 @@ SEXP argb32_to_raw_(SEXP argb32_) {
 
 
   SEXP result_ = PROTECT(allocVector(RAWSXP, w*h*4));
-  memcpy(RAW(result_), argb32, w*h*4);
+  memcpy(RAW(result_), packed_ptr, w*h*4);
 
   UNPROTECT(1);
   return result_;
@@ -70,16 +70,16 @@ SEXP argb32_to_raw_(SEXP argb32_) {
 
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-// Set dimensions on argb32
+// Set dimensions on packed pointer
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-SEXP argb32_set_dim_(SEXP argb32_, SEXP w_, SEXP h_) {
+SEXP packed_set_dim_(SEXP packed_ptr_, SEXP w_, SEXP h_) {
   SEXP outdim = PROTECT(allocVector(INTSXP, 2));
   INTEGER(outdim)[0] = INTEGER(h_)[0];
   INTEGER(outdim)[1] = INTEGER(w_)[0];
 
-  R_SetExternalPtrProtected(argb32_, outdim);
+  R_SetExternalPtrProtected(packed_ptr_, outdim);
   UNPROTECT(1);
-  return argb32_;
+  return packed_ptr_;
 }