+++ NEWS.md
@@ -1,3 +1,9 @@
+# pixelweaver 0.1.1
+
+* Make package a little more generic to support more pixel formats
+* Add support for: RGBA32, ABGR32
+
+
 # pixelweaver 0.1.0
 
 * Initial release